fix: database seeder
